[Chore] #134 - DateBoard/Report/Views ReportReasonView - 중복 코드 최소화, 팩토리 메소드 활용, Mark 추가
